% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omit_na.R
\name{omit_na}
\alias{omit_na}
\title{Removes rows with \code{NA} from a \code{data.frame}}
\usage{
omit_na(df)
}
\arguments{
\item{df}{A \code{data.frame}.}
}
\value{
Cleaned \code{data.frame} with rows that contained \code{NA} removed
}
\description{
Removes rows with \code{NA} from a \code{data.frame}
}
\details{
Removes rows with \code{NA} values from inputted \code{data.frame} and returns cleaned \code{data.frame}. Stores information about omitted rows in \code{na.action} attribute. Mimics the functionality of \code{na.omit()}, with a faster run time. This is useful for large data sets, or within functions that need to remove rows with \code{NA} values as a sub-step.
}
\examples{
omit_na(airquality)
omit_na(mtcars)

}
\seealso{
\code{\link[stats:na.fail]{stats::na.omit()}}
}
\author{
Emma Ruckle
}
